+++ src/x_info_generators/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.5.1"
+__version__ = "1.6.0"
 REPO_URL = "https://github.com/janroudaut/x-info-generators"